refactor: find-meeting 위젯 정리 및 FindMeetingPanel 컴포넌트 삭제
