+++ src/entities/meetings/model/hooks/useFilterCalendarStore.ts
@@ -1,14 +1,24 @@
 import { create } from 'zustand';
-import { MeetingFilters, SortType } from '../../types';
 
-interface Filters {
-  filters: MeetingFilters;
-  setFilters: (filter: MeetingFilters) => void;
+interface FilterState {
+  selectedCalendar: boolean;
+  toggleCalendarSelection: () => void;
+  date: Date;
 }
 
-const useFilterStore = create<Filters>()((set) => ({
-  filters: { sort: SortType.latest },
-  setFilters: (filter: MeetingFilters) => set({ filters: filter }),
+const useFilterCalendarStore = create<FilterState>()((set) => ({
+  selectedCalendar: false,
+  toggleCalendarSelection: () => {
+    set((state) => {
+      if (state.selectedCalendar) {
+        Object.assign(state, { selectedCalendar: false });
+      } else {
+        Object.assign(state, { selectedCalendar: true });
+      }
+      return { selectedCalendar: state.selectedCalendar };
+    });
+  },
+  date: new Date(),
 }));
 
-export default useFilterStore;
+export default useFilterCalendarStore;
